+++ dto/src/main/java/org/hotiver/dto/product/ProductProjection.java
@@ -1,11 +1,12 @@
 package org.hotiver.dto.product;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface ProductProjection {
     Long getId();
     String getName();
-    Double getPrice();
+    BigDecimal getPrice();
     String getDescription();
     String getCategoryName();
     String getImageUrl();
